backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.25.0
motor==3.3.1
pymongo==4.5.0
python-dotenv==1.0.0
python-multipart==0.0.22
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.0

sentence-transformers==3.0.1
faiss-cpu==1.13.2
scikit-learn==1.4.0
numpy==1.26.4
pandas==2.2.2
openpyxl==3.1.2
rank-bm25==0.2.2

groq==0.9.0
httpx==0.28.1

torch==2.2.0
transformers==4.38.0
huggingface-hub==0.20.0

requests==2.31.0
jinja2==3.1.3
dnspython==2.6.1
email-validator==2.1.1
click==8.1.7
typing-extensions==4.12.0


